+++ examples/real_time_enforcer/outputs.tf
@@ -16,65 +16,7 @@
 
 output "suffix" {
   description = "The random suffix appended to Forseti resources"
-  value       = "${module.forseti.suffix}"
-}
-
-#------------------------#
-# Forseti client outputs #
-#------------------------#
-
-output "forseti-client-vm-name" {
-  description = "Forseti Client VM name"
-  value       = "${module.forseti.forseti-client-vm-name}"
-}
-
-output "forseti-client-vm-ip" {
-  description = "Forseti Client VM private IP address"
-  value       = "${module.forseti.forseti-client-vm-ip}"
-}
-
-output "forseti-client-vm-public-ip" {
-  description = "Forseti Client VM public IP address"
-  value       = "${module.forseti.forseti-client-vm-public-ip}"
-}
-
-output "forseti-client-service-account" {
-  description = "Forseti Client service account"
-  value       = "${module.forseti.forseti-client-service-account}"
-}
-
-output "forseti-client-storage-bucket" {
-  description = "Forseti Client storage bucket"
-  value       = "${module.forseti.forseti-client-storage-bucket}"
-}
-
-#------------------------#
-# Forseti server outputs #
-#------------------------#
-
-output "forseti-server-vm-name" {
-  description = "Forseti Server VM name"
-  value       = "${module.forseti.forseti-server-vm-name}"
-}
-
-output "forseti-server-vm-ip" {
-  description = "Forseti Server VM private IP address"
-  value       = "${module.forseti.forseti-server-vm-ip}"
-}
-
-output "forseti-server-vm-public-ip" {
-  description = "Forseti Server VM public IP address"
-  value       = "${module.forseti.forseti-server-vm-public-ip}"
-}
-
-output "forseti-server-service-account" {
-  description = "Forseti Server service account"
-  value       = "${module.forseti.forseti-server-service-account}"
-}
-
-output "forseti-server-storage-bucket" {
-  description = "Forseti Server storage bucket"
-  value       = "${module.forseti.forseti-server-storage-bucket}"
+  value       = "${random_string.suffix.result}"
 }
 
 #--------------------------#