+++ test/integration/real_time_enforcer/controls/server.rb
@@ -1,1 +1,0 @@
-../../simple_example/controls/server.rb